+++ NEWS.md
@@ -1,3 +1,9 @@
+# b64 (development version)
+
+* Enables building in offline environments
+* Improves documentation
+* Adds new utility function `decode_as_string()` to decode base64 strings directly into a string.
+
 # b64 0.1.3
 
 * Addresses non API changes